@@ -27,8 +27,8 @@
   ],
   "type": "typo3-cms-extension",
   "require": {
-    "php": ">=8.1 <8.3",
-    "typo3/cms-core": ">=12.0 <12.9.99",
+    "php": "^8.2",
+    "typo3/cms-core": ">=13.4 <13.9.99",
     "google/cloud-storage": ">=v1.30 <1.31"
   },
   "autoload": {
