@@ -4,7 +4,7 @@ $EM_CONF[$_EXTKEY] = [
     'title' => 'Google Cloud Storage FAL Driver',
     'description' => 'Google Cloud Storage FAL driver for TYPO3. Files can be stored in the GCS buckets.',
     'category' => 'be',
-    'version' => '2.0.1',
+    'version' => '3.0.0',
     'state' => 'stable',
     'clearcacheonload' => 1,
     'author' => 'Pierre Geyer',
@@ -12,7 +12,7 @@ $EM_CONF[$_EXTKEY] = [
     'author_company' => 'next.motion',
     'constraints' => [
         'depends' => [
-            'typo3' => '12.0.0-12.9.99',
+            'typo3' => '13.4.0-13.9.99',
         ],
     ],
 ];
